Big changes I did most of this two years ago but neglected polishing it and presenting it until now. I hope you like it. The changes: * remove the requirement for lxml (to remove the requirement for a microsoft c++ redistributable) * remove the requirement for python-magic by being vague about image types * up the required version of pillow (to remove the requirement for zlib) * convert the main bash script into a file that does everything and is simultaneously a bash and batch script * moves the autogenerated cover to its own function, which is controlled by a global variable * includes, excludes, or collates authors notes depending on a global variable * replace the old book 1 image with a shorter one * add a gitignore to cache/ * use replacements.txt instead of manually replacing text in the program * nest the table of contents a little better * update readme
This update doesn't do anything about https://unsongbook.com/tosefta/ because I wasn't sure what to do with it.

I've tested this new code on windows and ubuntu on windows, so I think it works.
I haven't really tested the force feature very much because either I don't understand what it's supposed to do or it's just never worked for me. My hunch is that it was only useful when the book was actively updating?